Create a tokens page container
Created tokens page container that contains all tokens sub rotes.
